Eliminate code duplication in prev/next descriptors
