+++ core/optaplanner-core-impl/src/main/java/org/optaplanner/core/impl/domain/variable/nextprev/PreviousElementVariableListener.java
@@ -1,4 +1,4 @@
-package org.optaplanner.core.impl.domain.variable.previous;
+package org.optaplanner.core.impl.domain.variable.nextprev;
 
 import java.util.List;
